[JEWEL-821] Fix Jewel publishing
This adds a patchDependencies capability to MavenArtifactsBuilder, which
is used by JewelMavenArtifacts to change the dependencies that are added
to the POM files. This way, we set up the correct scopes and exclusions,
cleaning up unnecessary dependencies that only exist in the JPS build to
accommodate IJP quirks. This makes the artefacts usable.

closes https://github.com/JetBrains/intellij-community/pull/3056

GitOrigin-RevId: 5eb3adcbb16cd06352e720bde3b5b2ec3da47875