[tasks] convert intellij.platform.tasks.impl to a content module (IJPL-189531)
This is needed to have an explicit dependency on the backend in it (IJPL-188916), also, it fixes the problem with missing dependency on vcs.impl module (IJPL-160364).

The change shouldn't affect external plugins which use classes from that module, because the new module will be automatically added to their dependencies.

GitOrigin-RevId: 991b8cdc1547a9b44028e704dc3d5d576f53b7e9