IJPL-187340 Fix incorrect size adjustment of popups by toolbars
Toolbars update their containing popup or hint size when
the actions are updated. Normally it's a good thing to do,
but when fast track updates are suppressed and the toolbar
is shown over and over again, it creates redundant resizes
every time, increasing the size continuously.

Because fast track updates are only suppressed for toolbars
that are of a floating nature, they do not technically require
any size adjustments to fit. So we can safely disable size
adjustments for such toolbars.

This fix doesn't cover any cases when a toolbar
actually needs these adjustments, and fast track is still
suppressed. But there are no such toolbars at the moment.

The fix also doesn't cover the opposite case: when a floating
toolbar does not have fast track suppressed.
For such toolbars the proper fix would be to enable suppression.

GitOrigin-RevId: 5c0223179090ccc2c6a7af7f6422028aefd3efc9