IJPL-189640 Update getMinHeight() for NodeRenderer
The value hardcoded in SimpleColoredComponent.getMinHeight
was fully matching the default node height for trees in the
classic UI.

The updated value matches the default node height
for both UIs, all themes and Compact Mode.

The value is not changed in SimpleColoredComponent,
but overridden, because for other use cases different
value can be necessary. This change is for trees only.

GitOrigin-RevId: 620adda27c0e4dc8770a4258daa5629c887023aa