IJPL-189807 Fix File Structure popup top panel height
Quite a while ago its preferred height was set to be equal
to the header height. It worked fine for single-row layouts,
but for SQL there are more than 3 checkboxes,
so setting the preferred height to a single-row value
would squeeze the whole thing horribly.

Fix by only applying that height when there's a single row.
This is almost the minimum height now,
with the difference that it can still be reduced in height
if the layout manager wants it.

GitOrigin-RevId: 6ba5671fb317828ba31248c15ee6483a0d240afa