[JEWEL-829] Fix the bridge not reading the LaF
In JEWEL-558, a change was made to add a new flow to the combine in
SwingBridgeService. However, that new flow never emitted, causing
the whole combine to stall until it did. This in turn causes only
the theme defaults to be loaded, until the user changes the
"contrast scrollbars" option, making the faulty flow emit, and
unblocking the combine.

The fix is to ensure the flow emits on start, like all other flows
in the `combine` do.

closes https://github.com/JetBrains/intellij-community/pull/3062

GitOrigin-RevId: 4baebd4184055b58cc0084dd7b3316182624c2e1