[JEWEL-830] Fix Editor TextStyle line height
Swing is not using the font size when computing the base
lineHeight, that is then multiplied by the user-set factor
(default 1.2). Instead, they use FontMetrics.getHeight(),
which computes it as leading + ascent + descent.

This changes the Jewel line height computation for the
editor text style to use the same logic. It requires a
Graphics2D instance, which is obtained from a Swing
BufferedImage on demand, to obtain the same font metrics
as the Swing UI uses.

This gets us a much closer result, but is still not 100%
matching. It looks like this introduces a ~10% extra line
height in Compose compared to Swing. I will make a follow
up change when I figure out where in Skia's line height
usage there is a divergence from Swing's, but for now this
is already much better than what we did before.

This change also fixes the Jewel sample plugin.xml so it
doesn't crash at startup when run in the IJP build.

closes https://github.com/JetBrains/intellij-community/pull/3064

GitOrigin-RevId: e2cf47b2ee03437480242184011ae06a824f2e26