IJPL-61441 Fix the Copy action from floating toolbar popups
This is the first attempt, could really use some improvement.
The problem is that the floating toolbar is a lightweight hint,
and its popup is an implementation detail hidden inside.
So there's no direct way to go from a child popup to
the floating toolbar and then to its editor.

It also feels very out of place in AbstractPopup,
but because these popups are just regular action group
popups, there's no natural place to subclass / override
anything on the floating toolbar side.
(they invoked through simple button.click).

GitOrigin-RevId: b2ffded7ecfdccb13d102822d2a4ce8e6671b271