[java-dfa] Do not check nullability match for primitive methodref return types
DfaPsiUtil returns UNKNOWN nullability for primitive types, but in future we may return NOT_NULL, which is technically more correct. On the other hand, here implicit unboxing is possible, and it's already handled elsewhere, so we won't need an extra warning.
Part of IDEA-372347 Java type inference should respect nullability

GitOrigin-RevId: b864bfc4995a9b5e0e6a13cba807bd8b5b7d04ca