[jps build] don't add RuntimeModuleRepository from IDE's installation to JPS build classpath (IJPL-189949)
It's now taken from 'devkit.runtime.module.repository.jps' library together with devkit-runtime-module-repository implementation which uses it. This is needed to allow using a newer version of RuntimeModuleRepository which will support a more efficient format. When compiling older branches where the library doesn't include this artifact, it'll be added to the classpath by MonorepoJpsBuildProcessParametersProvider.

GitOrigin-RevId: 24612c87d38a3477bb80b83755cda86427e11a7a