[execution] IJPL-189943 Don't show a live indicator if a process was terminated before notification was started
ProcessHandler doesn't send any events before `startNotify()` is called and stores all pending events in a queue (com.intellij.execution.process.ProcessHandler.TasksRunner). When `startNotify()` is called, the event queue is executed first (because its listener was registered before everyone else in the constructor) and only then the `startNotified` event is dispatched.

As a result, we first receive the `processTerminated` event, and only then `startNotified`. And since the `processAdapter` relies on the fact that `startNotified` will be received first, we end up with an incorrect icon state.

GitOrigin-RevId: 99f6b95bc87629bd4e6397a182a3b8a5dd86e257