IJPL-188109 [terminal] Fix font size scaling issues
Yet another issue with font sizes: now that the terminal
has its own color scheme, the font size taken from there
is scaled by the editor, but we already put the scaled size
there. This leads to double scaling.

Fix by leaving the color scheme alone and always
having the unscaled size there. In case the font size
was changed in the terminal only, temporarily,
using pinch zoom and similar actions,
apply that font size to the editor directly.
This preserves the correct font size in the color scheme.

GitOrigin-RevId: 6980fca7a337c203dd47b50e57116e1f49062049