[kt, tests] Refactoring: remove versions from KotlinMppTestsContext
Expose 'testProperties' instead of fields like 'gradleVersion' in KotlinMppTestsContext.

Versions have a lot of complicated handling in KotlinMppTests (substitutions,
resolution from CI). This commit incapsulates all of that in KotlinTestProperties.

Why it's good? Because it makes `KotlinMppTestsContext` connected with KMP only
as much as KotlinTestProperties do (they are very KMP-aware at the moment, but
the next commit will change that)

GitOrigin-RevId: 2a7be6cf25895fe5d7d8ba3ff074f62e2ef8cbe0