[kt, tests] Refactoring: extract KMP/Gradle-unaware KotlinSyncTestsContext
This commit consists of three changes:
- simple rename "KotlinMppTestsContext" -> "KotlinSyncTestsContext"
- simple rename "KotlinMppTestsContextImpl" -> "KotlinMppTestsContext"
- one downcast in `buildGradleModel.kt`

The result is that all the common infra refers to KMP/Gradle-unaware
KotlinSyncTestsContext and KotlinTestProperties. There are a bunch of
places that know that they're working with KMP/Gradle (e.g.
AbstractKotlinMppGradleImportingTest) - they didn't even change
because they were already referring to KotlinMppTestsContextImpl, and
now just refer to a more appropriately named KotlinMppTestsContext.

This allows reusing the infra in Gradle-unaware subsystems (e.g.
Amper) with much less copy-paste. See the next commit for that

GitOrigin-RevId: 55c256dbba6e92f91cac25868db03572686a61e1