[vcs-log] clear references to storage on disposal to ensure that backing files can be deleted
Log uses memory-mapped files to store indexes and other references. When we try to clean up the files manually on index invalidation, it may so happen that someone is holding the reference to a disposed VcsLogManager, which in turn holds references to those mmpapped structures.
Those hard references can prevent the JVM from actually releasing file references, so files can't be deleted.

GitOrigin-RevId: 667d6e0e59b3344e95708edf1a673f3bd65edeae