IJPL-190348 Floating toolbar tab actions refactoring
Move both actions to a single file. Extract common base class
to reduce code duplication. Move companion object stuff
to private top-level functions, this fixes the "companion
object in extensions are bad" warning and generally
improves class loading performance.

Make classes internal instead of @Internal,
as they're not used in other modules.

This refactoring is a preparation for action promotion
needed to fix the issue with the conflict with the editor
Tab action, now that the editor is available in the data
context of the floating toolbar.

GitOrigin-RevId: 205d689785cbf693c8490880db2afbd96de430a4