[Jupyter Notebooks] Change GutterPainter to ExecutionCountHolder
The gutter painter is an obsolete functionality that is not used anymore. However, there is still some logic that uses the fields of gutter painter. The logic of execution count was extracted and the part about the gutter was removed

This was done during DS-7052


(cherry picked from commit 6c3ebf664d25eee608f8b1b26febdbc5f48415b5)

IJ-CR-167252

GitOrigin-RevId: de7d208e4b136558f2921e70e7ffd30bf70af2d2