[fleet] replace ktor server
Starting a kotlin server loads 1784 classes (815 of them are from kotlin/reflect) and takes 400ms. And we do this twice in two separate class loaders (or thrice if there is a supervisor, no one to blame but us here)
We use ktor server both for our http api (dock and workspace) and as an RPC transport.
- our http apis are simple rest requests that do not require all the elaborate "application" stuff that ktor provides
- there is no benefit in using websocket for loopback connections. while it is still easier to use network-backed transport in case we need to talk to a virtual machine/docker container/etc, http or tcp will do the job just fine.
- the fact that workspace http api and rpc endpoint are coupled as one
server is an oversight, one could exist without another in some
scenarios

This commit removes plugins/toolbox because it isn't used anywhere and
was in fact broken a long time ago

GitOrigin-RevId: 613c9dc271c1fad46e189cb015b9b68ed864b7b6