IJPL-189114 [debugger] Split XStackFrame.customizePresentation into three methods
One for text representation, another for making a UI component, and the old one. The latter allows something that wasn't possible before: to provide more than one presentation over time, making XDebuggerFramesList repaint when new presentations are available.

It also gives space to providing custom textual presentation of a frame, independent of what is shown in the UI. For example, in Java textual frames have a certain widely recognized format that differs from the UI representation in the IDE.

GitOrigin-RevId: 6a36309033228b8eae65aa20944ba0a27c65943c