IJPL-48531 Fix the DragRecognitionSupport leak
Due to the way it's designed, it may keep a reference
to some component even after the project is closed.
It's not accessible directly, but we can indirectly
clear its state by sending a "mouse released"
event to any component that supports this functionality.

For now, do it on the project frame close.
If the frame is not closed, then it's very likely
that the user clicks some component supporting DnD
inside the frame (say, Project View) and that will
clean up the reference. So any leaks that may
happen there are temporary.

GitOrigin-RevId: 39f278c75ac945e6283dc7c2998954bd77cea848