[runtime module repository] switch to use a more efficient format for the runtime module repository (IJPL-189949)
Now a module-descriptors.dat file is generated alongside module-descriptors.jar, and it's used to load the module descriptors at runtime. The new file contains the same data as module-descriptors.jar, but it's stored in binary format instead of XML files inside a ZIP file. This reduced the time required to load the data up to 10 times.

GitOrigin-RevId: 9139d637b12bc01085d7dcf03a273ad8779b9d49