IJPL-191230 RestoreSelectionListener: account for "loading..." nodes
When restoring selection after a tree update,
when a "loading..." node is replaced by the real nodes,
it wasn't selected because it didn't match the new node.
If node loading took long enough, and the user actually
selected the loading node, they normally expect
the selection to be kept on that node, not jump
to the parent.

Fix by considering the loading node to be matching
anything else. The first new node will match and stop
visiting there.

GitOrigin-RevId: ebde76e80be9473dc454319b1f076862d665080a