IJPL-189492 Fix backspace handling in YAML
After enabling backspace on both (backend and frontend) sides, it started to work incorrectly - removing more changes than it supposed to and causing care jumps.
The reason for that is backspaceModeOverride being registered on backend only, therefore backspace on frontend and backend behaves differently and causes caret jumps.

After moving backspaceModeOverride to the common module, backspace started to work fine

GitOrigin-RevId: ec15277ae4ecf2ccbd487572b62125bdf31817c6