IJPL-186705 Replace "left icons" with RowIcon
Because LEFT_ICONS needs a lot of special handling
in remdev (and the gap doesn't work at all),
replace it with a ready-to-use solution.

The gap is replaced by empty icons with the appropriate
width. It's not clear how to properly handle gap scaling
yet, but it's already better than gaps that don't work at all.

GitOrigin-RevId: e2ddfef26bc5479e6b0efef59983fa1c6d2cdb0d