[build] don't process product core plugins as plugins
plugin.xml files are usually considered a "plugin" while product platform plugin descriptor can also be named "plugin.xml". This can result in a situation when a product descriptor will be considered a plugin and processed accordingly. That is undesirable.
We know that the product descriptor plugin must always have "com.intellij" as an id, so we can try and filter out the descriptors this way.

GitOrigin-RevId: a15e7a2eef3c0e91eaf5bbfa66aeadca4a16f271