[project model] report an error if 'moduleConfigurationEditorProvider' extension requires 'Module' as constructor argument (IJPL-179175)
'moduleConfigurationEditorProvider' extension point was registered as a module-level. However, it seems that no implementations use this ability, and can be converted to application-level extensions automatically. To avoid breaking compatibility, the extension point was converted to 'bean class', and tries to call the constructor with 'Module' parameter if no-arg constructor isn't available, and report an error in such cases.

GitOrigin-RevId: 8390fda7aa09f09b6df8178d40b6269b26b0292f