IJPL-192020 Fix auto-sizing of the hidden toolbar actions popup
After disabling resize for toolbars with disabled fast track,
it now affects these popup toolbars that show the hidden actions.
It happens because SUPPRESS_FAST_TRACK is auto-set in
updateActionsImpl when firstTimeFastTrack is true.
For such toolbars fast track is useful to have
and auto-sizing is a must-have, so we just force
fast track by setting SUPPRESS_FAST_TRACK to false.

GitOrigin-RevId: d37eb67f6bbf506d347d76433f581706861e9107