[kotlin] fix the behavior of the "Setup SDK" link on the editor notification panel (KTIJ-34601)
Instead of using the deprecated 'chooseAndSetSdk' (see IJPL-190185), it now calls SdkPopupFactory directly. This fixes problems with the link doesn't work on the first click (because 'chooseAndSetSdk' always returns 'null' on the first call). Also, the popup is not shown not in the center of the IDE window, but underneath the link.

GitOrigin-RevId: 9c49eaf5482a8d5ce9e479b9eba5c1c4d53a1bfc