[spellchecker] cleanup: remove unused dependencies from intellij.spellchecker to XML modules
These dependencies aren't used in code, and they shouldn't there since intellij.spellchecker.xml was extracted to a separate module. Also, this change is needed to avoid failures in JetBrainsGatewayIdeStructureTest after 'intellij.spellchecker' is added to Gateway (to fix AT-2832).

Explicit dependency on intellij.platform.backend is added to ensure that the module isn't included in the frontend process (before dependency on intellij.platform.backend came as transitive).

GitOrigin-RevId: 75e55fda55205b8cc10ff08b93a7b2af1aefee43