IJPL-185425 [terminal] Ignore the first key typed event
The first event normally should never be a key typed event.
If we receive any other type of event, then dispatchKeyEvent
resets ignoreNextKeyTypedEvent to false.
But if the very first event was a key typed event,
it can only mean that the corresponding key pressed event
happened before the terminal was opened.
In this case we should just ignore it.

GitOrigin-RevId: 27499e6c3f1d976c5e1fbe98b118392030b5ce23