IJPL-187366 An attempt to fix race conditions in structure rebuild
The approach with concurrent launches is the main suspect
for various tool window glitches, like the content not showing
up or several tabs without names appearing in the header.
UI code is not designed to be concurrent, so when launching
several rebuilds, we need to wait until the previous one
has finished one way or another. But because of that
it seems that we don't even need concurrent launches.

To keep the cancel-on-content-manager-dispose semantics,
introduce a nested coroutine scope, cancelling it when
the tool window content manager is disposed.
Add some logging there as well so we can figure out
what's going wrong in case anything doesn't work.

GitOrigin-RevId: 743ce422c4c75d51635e7f459e68b83ec72d27c3