[remote dev] store the list of plugins disabled in the frontend process in a separate file (IJPL-192491)
The frontend process now stores the list of disabled plugins in a disabled_plugins_frontend.txt file instead of disabled_plugins.txt used by a regular IDE. When settings are imported from a previous version, the frontend imports data from disabled_plugins_frontend.txt if it is present, and otherwise imports disabled_plugins.txt to ensure that plugins which were disabled in a regular IDE will be disabled in the frontend process by default.

'installPerProcessInstanceSupportImplementation' calls are moved before 'PerProcessPathCustomization.prepareConfig' to ensure that P3Support is already initialized inside the latter.

GitOrigin-RevId: c60a152321d0d725d6c6f4ac249a00b4cfe9fb09