[git] IDEA-370108 Fix flaky DirectoryNotEmptyException in some tests
The root cause - `git add` reacting on VFS events can be executed after a project is disposed in parallel with folders cleanup, creating a race condition.
Such behavior is not pointing at a major production problem, so let's solve it on the per-test basis waiting for refresh to be performed.

GitOrigin-RevId: 7672672353481d795e682ca69c10b41e7ded2502