IJPL-189488 RD: Always fall back to the current editor in the Structure
For some reason the focused component on the backend
can be anything if a frontend-only component is focused.
It's not good when the Structure view can't update because
some irrelevant component is focused.
Let's fall back to the selected editor more aggressively.
It's better for the Structure view to follow the selected
editor than to simply stop updating.

GitOrigin-RevId: 25fe0a7e27c78eeba230abb2c3e39b4177e0bd38