IJ-CR-165937 Migrate CompositeFilterWrapper to flows
Using the usual flow-of-requests pattern ensures
sequential computations (no races between async
computations) without manual job housekeeping.

Update the tests to cancel the coroutine scope
explicitly, because the collector is now always-running,
the tests would hang and time-out otherwise.

To make awaitFiltersComputed() work reliably,
we had to add a Boolean flag to the result flow
so that tests can wait until the value is not only
published, but the listeners have been fired as well.

GitOrigin-RevId: 6696b2775f25b27a75398b46575fd73d72963c61