IDEA-373402 String#formatted() parameter highlighting is getting stuck when switching between parameters
do not cache the entire string literal range because there might be many sub-symbols inside this string literal for which the find usages are computed separately.
E.g. for formatted string modifiers "%s %d" we compute the usages of "%s" and "%d" separately.
So when the caret moves on the string literal, there are many identifier highlighting results possible.

GitOrigin-RevId: 63df5752c9a3fcc7f543008568da9484cbed5872