[plugin model] process dependencies on modules in 'processAllNonOptionalDependencies' (IJPL-192427, IJPL-192738)
Before, only dependencies on plugins were processed. So if a plugin depends on a module from another plugin, such dependencies weren't handled in Settings | Plugins, for example. Now `processAllNonOptionalDependencies` processes dependencies on modules as well. To find a descriptor by the module ID (name), a corresponding mapping is built by `PluginSet::buildContentModuleIdMap` and propagated to `processAllNonOptionalDependencies` along with `pluginIdMap`.

'getNonOptionalDependenciesIds' is deprecated since it doesn't return dependencies on modules. It's now used only in the performance testing plugin, so this shouldn't cause problems.

GitOrigin-RevId: 5ea349497169077cfe33cea4895ce853600fe856