[RDCT] Add `BackendProjectAwareDocumentExtensionsProvider` API
Documents are almost always used in the context of some projects, especially when they have some language support. `Document` by itself is an application-level entity, the same as `Backend/FrontendDocumentHost` is. But the document binding happens as a result of binding some editor, which can be associated with a project. This allows us to collect some (project + document) information at this stage and attach it to the document extensions.

One notable example is language support. Having a project for a document allows getting a corresponding `PsiFile`, which is crucial for the language support.

This commit adds a project-aware document extensions EP, which allows to collect early information related to a document and its language support.


(cherry picked from commit 5abfe7d09db4215a56be0141f2ce5d92f4dd81d5)

IJ-CR-166874

GitOrigin-RevId: d0e5a6413938df371c5e45d978b381aabefaa5c1