IJPL-192256 Fix the window header not entering the compact state
Now that the Trial State Widget can't be removed
from the main toolbar (it's a NonCustomizableAction),
the check for the toolbar being empty needs to be fixed.
We can't just use the NonCustomizableAction in this check
because when the header is initialized, the actions may
not be loaded yet, and action stubs don't implement
the marker interface. So we have to resort to an ID-based
check here, and we have to hardcode the ID because
it's not exposed anywhere.


(cherry picked from commit debb3f7c464220d050a263e7e7f864245d0ac961)

IJ-CR-166609

GitOrigin-RevId: 5ff7b4647605aa865d76f1eb1f3d121e74416175