[rdct] IJPL-193192: Use already shown skeleton's data for a new skeleton when possible
In the case of switching tabs, where two tabs and loading simultaneously, each tab had their own timer for the skeleton's delay. That means that if the tab had been switched, it would start a new timer and will show an empty tab even if the skeleton had been showing in the previous. This commit introduces an internal API and a way of stealing animation details and instantly spawn a skeleton whenever possible

GitOrigin-RevId: 572a58ee5833c443cb02b7416c55b682f8df3926