[gradle] IDEA-370426 add GradleLightProject.getIdentityPath
The `identityPath` is a unique path that identifies a project within the entire build hierarchy, including any included builds. It is constructed relatively to the root settings.gradle(.kts) file. Examples:
- `:`, for a root project of a root build
- `:anySubproject`, for a subproject of the root build
- `:anyIncludedBuild`, for a root project of an included build
- `:anyIncludedBuild:anySubproject`, for a subproject of an included build

In another words, `identityPath` is a prefix written before a task name to run the task of specific project in the build hierarchy.

Starting from Gradle 8.2, there is the `BasicGradleProject.getBuildTreePath` method that returns the same value as `identityPath`.

Tests:
- org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test DefaultGradleLightProject.getProjectIdentityPath
- org.jetbrains.plugins.gradle.model.DefaultGradleLightBuildTest.test DefaultGradleLightBuild.convertGradleBuilds

GitOrigin-RevId: a973e505868a129bd404e3b13ff43ade4c85a7e2