IJPL-193277 The Structure View: retrieve the editor from the data context
The previous code was seemingly correct, but for some reason
didn't always work, sometimes returning an outdated editor.
Anyway, it's not recommended to directly retrieve anything
about the current state of the UI after the data context
was captured, as it may lead to inconsistent results.
The whole point of the data context is to avoid inconsistencies
by having a consistent snapshot of the UI.

Use the LAST_ACTIVE_FILE_EDITOR key as it's known
to be the most reliable one for retrieving the current editor.

GitOrigin-RevId: 6582879a7bb7ffb9d819b29dad804740e103fe80