[terminal] IJPL-193309 Migrate TerminalTypeAhead to the blocks model
This simplifies the code and fixes the bug with storing
the index that may get outdated later.

Because of the extra dependency, the code that creates
everything in ReworkedTerminalView had to be rewritten
to pass the type ahead instance as user data into the editor.
Otherwise, there would be a cyclic dependency,
as the instance is required to configure the editor,
and the editor is required to configure the blocks model.

GitOrigin-RevId: df9436fc82b721a90efc59a9673204201bb425b6