[terminal] IJPL-192022 Create separate internal method to create terminal tab for the selected Terminal Engine
And use it in the platform places that should create the terminal tab with the Terminal Engine specified in the settings.
In all other places (in internal and external plugins) the classic terminal will be used regardless of the engine specified in the settings.
This way we will preserve the compatibility with existing terminal tab creation API usages.

(cherry picked from commit 94477d864c5c43750ccfbb32ffc9e1d5bfc717d4)

IJ-CR-167039

GitOrigin-RevId: 3ebeff21e3ec958aa93a4f14754e6d22639de069