[rdct] IJPL-192929: Fetch the correct file id if the file is not found by id
After VFS cache invalidation, restoring from the frontend will reportedly fail. The reason is reset of file IDs. That means that there will be no more correct file ids on the backend, thus the requested file could not be found. This commit adds the ability to fetch correct id if the incorrect has been supplied from the editor + a registry flag to test it

GitOrigin-RevId: 34372d65582728e0c791687a14eb62b186e5b194