IJPL-187934 Fix the Switcher preferred size
The Switcher preferred size used to be computed
in the monolith using the number of files,
through the visibleRowCount property.
It was lost at some point when splitting the feature.
Bring it back.

There were some other UI code next to this line
that was also lost, that part is to be investigated
and discussed. But this line alone fixes the regression.

GitOrigin-RevId: 61eaa52e0742f6c450d35a2fdfe65cfa69b2894c