IJPL-187934 Fix the Switcher preferred size
The Switcher preferred size used to be computed
in the monolith using the number of files,
through the visibleRowCount property.
It was lost at some point when splitting the feature.
Bring it back.

There were some other UI code next to this line
that was also lost, that part is to be investigated
and discussed. But this line alone fixes the regression.


(cherry picked from commit 61eaa52e0742f6c450d35a2fdfe65cfa69b2894c)

IJ-CR-167210

GitOrigin-RevId: 8a16f3718a217db8583e22d9ae3cff29e715ac9f