IJPL-188608 [debugger] Ignore stackFrameChanged event from backend if initiated by frontend
On the frontend side, currentStackFrame is updated optimistically: frontend session doesn't wait for a roundtrip to the backend,
and neither do all the frontend views. Once the frame is changed in the UI, the frontend considers it the current stack frame.

The backend receives this frame later, then updates its structures, then broadcasts this event across its listeners
(see com.intellij.xdebugger.impl.XDebugSessionImpl.setCurrentStackFrame).

The frontend has to listen to the broadcaster because the frame can be changed outside the frames view
(for example, via Show Execution Point, and theoretically by anyone who wants to change the frame)
i.e., it could be the frame of the second type.

In an ideal world, when there's no latency between FE and BE, the frame coming from the backend is the same
as the current stack frame on the frontend.

However, if there's latency, and the frames are changing fast in the UI, the following can happen:
  - frame f1 is selected
  - frontend's current stack frame is now f1
  - frontend sends f1 to the backend
  - frame f2 is selected
  - frontend's current stack frame is now f2
  - frontend sends f2 to the backend
  - backend broadcasts f1 to the frontend
  - (*) frontend's current stack frame is now f1
  - frontend sends f1 to the backend
  - ...

So, there is a potentially endless update cycle (see IJPL-188608 for fun outcomes).

The problem is that we cannot just ignore the event on step (*), because it might be the frame of the second type,
and the frontend has to update its own currents.

This change helps discern request to set stack frame initiated by frontend and those initiated by backend by adding an overload to stackFrameChanged to the dispatcher.

GitOrigin-RevId: 4a59c02ffaf247519eaa92b1c1124c799c734f7b