[terminal, feedback] IJPL-194633 Introduce suspending system data API
To allow accessing TerminalProjectOptionsProvider.getShellPath
when computing the system data for the feedback,
introduce a new suspending API and port all usages to it.

This might look like an overkill just for the terminal,
but using blocking I/O, OS API and whatever else
looks like a common case for collecting system info data
in general, so it's nice to have anyway.

Because all this info is collected for a dialog,
using a coroutine bound to that dialog sounds like a good
place for the computation.
To access the result on the EDT when needed,
the result is awaited under a modal progress.
But in order for it to be quick, the computation is started
when the dialog is shown.
So in most cases it won't even be necessary to await it.

For the AIA specifically, we introduce an API that allows
to access the computation result on any thread,
so it can be changed (it adds an upload ID to it).
To ensure the computation result is available after the dialog
is closed, we also await the result explicitly
in doOKAction().

GitOrigin-RevId: 2edb1baf10e508c801c32427cd09d4473e44552b