[Plugin Advertisement] Allow dependency collectors to define a custom reason why the plugin is suggested
The customization for the suggestion reason already existed: for detected CLI tools there was a different suggestion reason, rather then to the detected libraries of the project.

However, this customization was hard-coded. With this update, it can now be defined by the collector.

This update is needed for GO-17806 where a different reason for suggestion will be defined

(cherry picked from commit be9fbeb2d9dfafc18a9ea38ca21fee56fec8e6ec)

GitOrigin-RevId: 0abd1a8b225b3a5aeaa02fe94c78958f1ef97a13