[Plugin Advertisement] Allow dependency collectors to define a custom reason why the plugin is suggested
The customization for the suggestion reason already existed: for detected CLI tools there was a different suggestion reason, rather then to the detected libraries of the project.

However, this customization was hard-coded. With this update, it can now be defined by the collector.

This update is needed for GO-17806 where a different reason for suggestion will be defined

GitOrigin-RevId: be9fbeb2d9dfafc18a9ea38ca21fee56fec8e6ec