support EffectType.FADED in EditorPainter
TextAttributes have an EffectType called FADED. It makes text
semi-transparent. It was added by 5cd5531808a43e29f5e8469b48d35387b838828c
and it is currently supported in the project tree, but is ignored
in the editor. For CLion and for Rider for GameDev I want to make
text semi-transparent. At the moment, I want to highlight inactive
preprocessor branches, but in the future, it may be possible to
highlight unreachable code in the same way. This way, we will
match the bahavior of Visual Studio and of ReSharper.

GitOrigin-RevId: 114abb08421ed07fb83c7ab3331ba1ef63cfff96