[rdct] IJPL-194924: Do not even try to open files from previous VFS
If the file with fileId exceeding the current maximumFsRecord is tried to be opened it:
* Will fail
* Will start to throw a lot of exceptions about invalid and corrupted id.

That leads to the unability to reopen files after 252->253 migration in some cases.

The fix introduces a new field for `RdFileIdNew` — `managingFsModificationTimestamp` which states backend's `ManagingFS.getCreationTimestamp()`. That allows to fail fast on the outdated files and do not corrupt new VFS

GitOrigin-RevId: 61138289ed3d3342e11b3b4b7f26cd2b855d007a