[terminal] IJPL-194863 Update pty4j to 0.13.9
This version reverts the ConPTY update that caused
the regression on Windows 10.

Reviewed and approved in IJ-CR-167781.
The changes are the same as in the original
commit except that the file lib-lock.json
doesn't exist in 252, hence the manual cherry-pick.

(cherry picked from commit c9d9e38a6e6ff194120788a90977af9f2a43053f)

GitOrigin-RevId: 5a17bfac2f2d5fadf60398cb96c3555c78992c24