PY-79816 Introduce `PyType#getMemberTypes` and use it to infer `__hash__` type of dataclasses
`getMemberTypes` should be used for members which have no PSI which can be used to resolve to. For example, `__init__` method in dataclasses are sometimes not present in the source code. Yet its parameters are always useful for the code analysis. In this case, `getMemberTypes` should be used.

(cherry picked from commit 2455ed05099842fc50e1fa2a196c4952b6444795)

IJ-CR-172940

GitOrigin-RevId: d68e9e7513727ece49a796f29e241ee921f223e6