IJPL-194248 Fix race condition causing freeze on startup
Editors are opened in parallel during startup.
It can happen that `startupFileEntry` is not set before the next file is created.

In such cases, the file goes through the cache to look for an already created `VirtualFile` instance.
We compare files using the `path`, which we obtain from either `fileProperties` or `startupFileEntry`.
If neither is set, we wait for `fileProperties` to be received from the backend (with 3s timeout), which can cause a freeze.

(cherry picked from commit 8f266311492057c4061281cc31e8752850d615b8)

GitOrigin-RevId: 37e94c0abbcd9fc869c38ed0a10b0531f449c34e