IJPL-194248 Fix race condition causing freeze on startup
Editors are opened in parallel during startup.
It can happen that `startupFileEntry` is not set before the next file is created.

In such cases, the file goes through the cache to look for an already created `VirtualFile` instance.
We compare files using the `path`, which we obtain from either `fileProperties` or `startupFileEntry`.
If neither is set, we wait for `fileProperties` to be received from the backend (with 3s timeout), which can cause a freeze.

(cherry picked from commit 8f266311492057c4061281cc31e8752850d615b8)


(cherry picked from commit 1dbdb97a5952cda0cbc88eb17ccaace881620485)

IJ-MR-167365

GitOrigin-RevId: e0fd1bfe3a5bfbdb622add41b8f69c3b550363fe