[gradle] refactor DefaultGradleLightBuildTest
- Split the single complex test into two separate tests: for the simple case and for a composite build.
- Simplify mocking and verification. Clarify what project structure is created and what values are expected. It makes the test longer, however, simplifying it improves its reliability.
- Add messages for assertions
- Fix `buildTreePath`

The test did specify the wrong `buildTreePath` for project mocks. It didn't fail because for Gradle versions >= 8.2, the `identityPath` is taken from `buildTreePath`, and the test was running only for a fresh Gradle version. So, identityPath was equal to the wrongly specified `buildTreePath`.

For Gradle versions < 8.2, the buildTreePath is calculated in `DefaultGradleLightProject#getProjectIdentityPath`. This behavior is already checked in another test: `.DefaultGradleLightProjectTest.test DefaultGradleLightProject.getProjectIdentityPath`. However, since the mistake was made in `buildTreePath`, it makes sense to verify the `identityPath` also here for different Gradle versions.

Tests:
- org.jetbrains.plugins.gradle.model.DefaultGradleLightBuildTest.test DefaultGradleLightBuild.convertGradleBuilds for simple build
- org.jetbrains.plugins.gradle.model.DefaultGradleLightBuildTest.test DefaultGradleLightBuild.convertGradleBuilds for composite build

Is done as a previous step for IDEA-370426, but is far from the context described in the ticket.

GitOrigin-RevId: 685f032fd53f9b555dea7110340bff7efb9cb441