IJPL-184315 Move `intellij.sh.shellcheck` to backend
The module runs the external shellcheck tool (downloads it, reads the files) to analyze shell scripts and provide annotations in the editor.
Annotator (`ShShellcheckExternalAnnotator`) is moved to the backend because it executes external shellcheck processes, performs PSI analysis and interacts with the file system. Inspection (`ShShellcheckInspection`) is moved to the backend, to avoid duplication. Intentions interact with backend features, so they are moved to the backend.
The `ShellcheckOptionsPanel` is kept in the core, because it's a UI component that provides a configuration interface directly used by users to set up shellcheck settings. Similar with `ShellcheckSetupNotificationProvider` and util class, used by UI components.

GitOrigin-RevId: 0b0f5bedf77cfd7a2de1d1ca6ad3b6a288f44967